update file structure, move certain functions to be shared, address comments
