+++ Demo/Demo/Extensions/UIViewController+Extension.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+extension UIViewController {
+
+    func removeChildViews() {
+        children.forEach { viewController in
+            viewController.willMove(toParent: nil)
+            viewController.view.removeFromSuperview()
+            viewController.removeFromParent()
+        }
+    }
+}
